requirements.txt
django
djangorestframework
psycopg2
pillow
python-dotenv
django-filter
djangorestframework-simplejwt
